+++ spec/l3_agent_spec.rb
@@ -40,7 +40,7 @@ describe 'openstack-network::l3_agent' do
         expect(chef_run).to create_template(file.name).with(
           user: 'neutron',
           group: 'neutron',
-          mode: 0644
+          mode: 0640
         )
       end