+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of cookbook-openstack-net
 ## 10.0.0
 * Upgrading to Juno
 * Sync conf files with Juno
+* Upgrading berkshelf from 2.0.18 to 3.1.5
 
 ## 9.1.1
 * Allow dhcp_delete_namespaces and router_dhcp_namespaces to be overridden.